My version for modeling MAR melting
I have played with the main function PlotPseudosection.mlapp to
learn. Finally, I save the model results for N-MORB with higher
and lower P-T resolution, which makes minor differences, so the
low-resolution version is good enough. These mat files will be
used later in the MAR modeling in the parental folder.

	modified:   EquilibriumPath_MAGEMin.mlx
	new file:   MAGEMin-N-MORB-lowRes.mat
	new file:   MAGEMin-N-MORB.mat
	modified:   PlotPseudosection.mlapp
	modified:   matlab/Plot_IsoContour.m